UEFI: barebone bootloader completed
