+++ xv7-bootloader-uefi/src/mem.rs
@@ -1,6 +1,7 @@
 use uefi::prelude::*;
 use uefi::Result;
 
+#[allow(unused)]
 pub fn memory_map(services: &BootServices) -> Result<()> {
     let map_size = services.memory_map_size();
     let mut buf = vec![0u8; map_size];
